HW1/requirements.txt
numpy==1.21.4
paho_mqtt==1.6.1
requests==2.22.0
